requirements.txt
boto3==1.42.0
botocore==1.41.6
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
urllib3==2.5.0
